requirements.txt
agno[all]
agno[xai]
python-dotenv
pydantic
sqlalchemy
fastapi
uvicorn[standard]
httpx